organized remote generation code
