--- testing/project.chor
@@ -1,38 +1,0 @@
-{
- "name":"project",
- "version":"v2025.0.0",
- "variables":{
-  "expressions":{
-   "name":{
-    "unit":"Meter",
-    "var":["pose.x()",0.0]
-   }
-  },
-  "poses":{
-   "pose":{
-    "x":["0 m",0.0],
-    "y":["0 m",0.0],
-    "heading":["0 rad",0.0]
-   }
-  }
- },
- "config":{
-  "modules":[
-    {"x":["11 in",0.2794], "y":["11 in",0.2794]},
-    {"x":["-11 in",-0.2794], "y":["11 in",0.2794]},
-    {"x":["-11 in",-0.2794], "y":["-11 in",-0.2794]},
-    {"x":["11 in",0.2794], "y":["-11 in",-0.2794]}],
-  "mass":["150 lbs",68.0388555],
-  "inertia":["6 kg m ^ 2",6.0],
-  "gearing":["6.5",6.5],
-  "radius":["2 in",0.0508],
-  "vmax":["6000 RPM",628.3185307179587],
-  "tmax":["1.2 N * m",1.2],
-  "bumper":{
-   "front":["16 in",0.4064],
-   "left":["16 in",0.4064],
-   "back":["16 in",0.4064],
-   "right":["16 in",0.4064]
-  }
- }
-}
